etl/requirements.txt
psycopg2-binary==2.9.3
pandas==1.1.5
python-dotenv==0.16.0
